+++ tests/conftest.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 from posixpath import join as urljoin
 
 from requests import HTTPError
-from six.moves.urllib.parse import urlencode, quote
+from urllib.parse import urlencode, quote
 from mock import Mock
 
 from airtable import Airtable