+++ HISTORY.md
@@ -1,5 +1,6 @@
 # 0.15.0
 * **Breaking**: Drop Api config from ENV variable - use `api_key` arg instead
+* **Breaking**: Drop Python 2 / IronPython Support
 
 
 # 0.14.0